tests in own mod with root_device consts & helpers
